+++ src/index.html
@@ -15,7 +15,7 @@
 <body>
   <div class="page">
     <!-- Попап редактирования аватара -->
-    <div class="popup popup_opened" id="edit-ava">
+    <div class="popup" id="edit-ava">
       <div class="popup__container">
         <button class="popup__close-icon" type="button"></button>
         <form action="#" method="POST" name="profileAvatar" id="profileAvatar" class="form" novalidate>
@@ -30,7 +30,7 @@
       </div>
     </div>
     <!-- Попап редактирования профиля -->
-    <div class="popup" id="edit">
+    <div class="popup popup_opened" id="edit">
       <div class="popup__container">
         <button class="popup__close-icon" type="button"></button>
         <form action="#" method="POST" name="profileInfo" id="profileInfo" class="form" novalidate>
